[Fix]: Rename utils/typing to utils/typing_utils to fix collect_env (#9265)
