+++ mmdet/models/dense_heads/rtmdet_head.py
@@ -10,8 +10,7 @@ from torch import Tensor
 
 from mmdet.registry import MODELS, TASK_UTILS
 from mmdet.structures.bbox import distance2bbox
-from mmdet.utils import reduce_mean
-from mmdet.utils.typing import ConfigType, InstanceList, OptInstanceList
+from mmdet.utils import ConfigType, InstanceList, OptInstanceList, reduce_mean
 from ..layers.transformer import inverse_sigmoid
 from ..task_modules import anchor_inside_flags
 from ..utils import (images_to_levels, multi_apply, sigmoid_geometric_mean,
