+++ mmdet/models/layers/csp_layer.py
@@ -5,7 +5,7 @@ from mmcv.cnn import ConvModule, DepthwiseSeparableConvModule
 from mmengine.model import BaseModule
 from torch import Tensor
 
-from mmdet.utils.typing import ConfigType, OptConfigType, OptMultiConfig
+from mmdet.utils import ConfigType, OptConfigType, OptMultiConfig
 from .se_layer import ChannelAttention
