+++ mmdet/models/roi_heads/bbox_heads/sabl_head.py
@@ -6,6 +6,8 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from mmcv.cnn import ConvModule
+from mmengine.config import ConfigDict
+from mmengine.structures import InstanceData
 from torch import Tensor
 
 from mmdet.models.layers import multiclass_nms
@@ -13,8 +15,7 @@ from mmdet.models.losses import accuracy
 from mmdet.models.task_modules import SamplingResult
 from mmdet.models.utils import multi_apply
 from mmdet.registry import MODELS, TASK_UTILS
-from mmdet.utils.typing import (ConfigDict, ConfigType, InstanceData,
-                                InstanceList, OptConfigType, OptMultiConfig)
+from mmdet.utils import ConfigType, InstanceList, OptConfigType, OptMultiConfig
 from .bbox_head import BBoxHead
