+++ mmdet/utils/__init__.py
@@ -10,9 +10,9 @@ from .misc import (find_latest_checkpoint, get_test_pipeline_cfg,
 from .replace_cfg_vals import replace_cfg_vals
 from .setup_env import register_all_modules, setup_multi_processes
 from .split_batch import split_batch
-from .typing import (ConfigType, InstanceList, MultiConfig, OptConfigType,
-                     OptInstanceList, OptMultiConfig, OptPixelList, PixelList,
-                     RangeType)
+from .typing_utils import (ConfigType, InstanceList, MultiConfig,
+                           OptConfigType, OptInstanceList, OptMultiConfig,
+                           OptPixelList, PixelList, RangeType)
 
 __all__ = [
     'collect_env', 'find_latest_checkpoint', 'update_data_root',
